Got the trending API call working

+++ EquityX/APIResponse/QuoteResponse/QuoteResponse.cs
@@ -17,8 +17,8 @@
         public string fullExchangeName { get; set; }
         public string longName { get; set; }
         public string financialCurrency { get; set; }
-        public int averageDailyVolume3Month { get; set; }
-        public int averageDailyVolume10Day { get; set; }
+        public long averageDailyVolume3Month { get; set; }
+        public long averageDailyVolume10Day { get; set; }
         public decimal fiftyTwoWeekLowChange { get; set; }
         public decimal fiftyTwoWeekLowChangePercent { get; set; }
         public string fiftyTwoWeekRange { get; set; }
@@ -26,7 +26,7 @@
         public decimal fiftyTwoWeekHighChangePercent { get; set; }
         public decimal fiftyTwoWeekLow { get; set; }
         public decimal fiftyTwoWeekHigh { get; set; }
-        public int dividendDate { get; set; }
+        public long dividendDate { get; set; }
         public decimal bookValue { get; set; }
         public decimal fiftyDayAverage { get; set; }
         public decimal fiftyDayAverageChange { get; set; }
@@ -37,10 +37,10 @@
         public object marketCap { get; set; }
         public decimal forwardPE { get; set; }
         public decimal priceToBook { get; set; }
-        public int sourceInterval { get; set; }
+        public long sourceInterval { get; set; }
         public string exchangeTimezoneName { get; set; }
         public string exchangeTimezoneShortName { get; set; }
-        public int exchangeDataDelayedBy { get; set; }
+        public long exchangeDataDelayedBy { get; set; }
         public string market { get; set; }
         public string shortName { get; set; }
         public decimal preMarketPrice { get; set; }
@@ -53,9 +53,9 @@
         public int askSize { get; set; }
         public string messageBoardId { get; set; }
         public string marketState { get; set; }
-        public int earningsTimestamp { get; set; }
-        public int earningsTimestampStart { get; set; }
-        public int earningsTimestampEnd { get; set; }
+        public long earningsTimestamp { get; set; }
+        public long earningsTimestampStart { get; set; }
+        public long earningsTimestampEnd { get; set; }
         public decimal trailingAnnualDividendRate { get; set; }
         public decimal trailingPE { get; set; }
         public decimal trailingAnnualDividendYield { get; set; }
@@ -66,19 +66,19 @@
         public object sharesOutstanding { get; set; }
         public bool esgPopulated { get; set; }
         public bool tradeable { get; set; }
-        public int gmtOffSetMilliseconds { get; set; }
+        public long gmtOffSetMilliseconds { get; set; }
         public int priceHint { get; set; }
         public decimal preMarketChange { get; set; }
         public decimal preMarketChangePercent { get; set; }
-        public int preMarketTime { get; set; }
+        public long preMarketTime { get; set; }
         public string exchange { get; set; }
         public decimal regularMarketPrice { get; set; }
-        public int regularMarketTime { get; set; }
+        public long regularMarketTime { get; set; }
         public decimal regularMarketChange { get; set; }
         public decimal regularMarketOpen { get; set; }
         public decimal regularMarketDayHigh { get; set; }
         public decimal regularMarketDayLow { get; set; }
-        public int regularMarketVolume { get; set; }
+        public long regularMarketVolume { get; set; }
         public string symbol { get; set; }
     }
